[Members] Move Tasks to own file, rename files in kebab-case
