+++ src/pages/members/member-info-block.js
